Update Random method
You may now use '-?' to get  a list of some movies in a random order, with possible duplicates but NOT in sequence. 

Just like your music player: if you have hundreds of songs and let it shuffle, you might want to listen to the same song before your entire playlist is played; you'd also be bored if the same one was played in sequence. I used the same principle here.